starting Train class integration into main

+++ Segmentation/utils/losses.py
@@ -1,4 +1,4 @@
-import numpy as np
+# import numpy as np
 from tensorflow.keras.losses import binary_crossentropy, categorical_crossentropy
 import tensorflow.keras.backend as K
 import tensorflow as tf
@@ -6,8 +6,10 @@ import tensorflow as tf
 epsilon = 1e-5
 smooth = 1
 
+
 def dice_coef_loss(y_true, y_pred):
     return -dice_coef(y_true, y_pred)
+
 
 def dice_coef(y_true, y_pred, smooth=1):
     y_true_f = K.flatten(y_true)
@@ -15,6 +17,7 @@ def dice_coef(y_true, y_pred, smooth=1):
 
     intersection = K.sum(y_true_f * y_pred_f)
     return (2. * intersection + smooth) / (K.sum(y_true_f * y_true_f) + K.sum(y_pred_f * y_pred_f) + smooth)
+
 
 def dsc(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
@@ -25,10 +28,12 @@ def dsc(y_true, y_pred):
     score = (2. * intersection + smooth) / (K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
     return score
 
+
 def dice_loss(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
     loss = 1 - dsc(y_true, y_pred)
     return loss
+
 
 def tversky_loss(y_true, y_pred, alpha=0.5, beta=0.5, smooth=1e-10):
     """ Tversky loss function.
@@ -52,6 +57,7 @@ def tversky_loss(y_true, y_pred, alpha=0.5, beta=0.5, smooth=1e-10):
 
     return 1 - answer
 
+
 def tversky_crossentropy(y_true, y_pred):
 
     tversky = tversky_loss(y_true, y_pred)
@@ -59,6 +65,7 @@ def tversky_crossentropy(y_true, y_pred):
     crossentropy = K.mean(crossentropy)
 
     return tversky + crossentropy
+
 
 def iou_loss(y_true, y_pred, smooth=1):
     y_true = K.flatten(y_true)
@@ -69,15 +76,18 @@ def iou_loss(y_true, y_pred, smooth=1):
 
     return iou
 
+
 def bce_dice_loss(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
     loss = binary_crossentropy(y_true, y_pred) + dice_coef_loss(y_true, y_pred)
     return loss
 
+
 def cce_dice_loss(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
     loss = categorical_crossentropy(y_true, y_pred) + dice_coef_loss(y_true, y_pred)
     return loss
+
 
 def iou_loss_eval_3d(y_true, y_pred):
     y_true = tf.slice(y_true, [0, 0, 0, 0, 1], [-1, -1, -1, -1, 6])
@@ -85,11 +95,13 @@ def iou_loss_eval_3d(y_true, y_pred):
     iou = iou_loss(y_true, y_pred)
     return iou
 
+
 def dice_coef_eval_3d(y_true, y_pred):
     y_true = tf.slice(y_true, [0, 0, 0, 0, 1], [-1, -1, -1, -1, 6])
     y_pred = tf.slice(y_pred, [0, 0, 0, 0, 1], [-1, -1, -1, -1, 6])
     dice = dsc(y_true, y_pred)
     return dice
+
 
 def dice_loss_weighted_3d(y_true, y_pred):
     dice_whole = 1 - dsc(y_true, y_pred)
@@ -97,6 +109,7 @@ def dice_loss_weighted_3d(y_true, y_pred):
     y_pred = tf.slice(y_pred, [0, 0, 0, 0, 1], [-1, -1, -1, -1, 6])
     dice = 1 - dsc(y_true, y_pred)
     return dice + dice_whole
+
 
 def precision(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
@@ -110,6 +123,7 @@ def precision(y_true, y_pred):
     prec = (tp + smooth) / (tp + fp + smooth)
     return prec
 
+
 def recall(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
     smooth = 1
@@ -121,6 +135,7 @@ def recall(y_true, y_pred):
     fn = K.sum(y_pos * y_pred_neg)
     recall = (tp + smooth) / (tp + fn + smooth)
     return recall
+
 
 def confusion(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
@@ -136,6 +151,7 @@ def confusion(y_true, y_pred):
     recall = (tp + smooth) / (tp + fn + smooth)
     return prec, recall
 
+
 def tp(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
     smooth = 1
@@ -143,6 +159,7 @@ def tp(y_true, y_pred):
     y_pos = K.round(K.clip(y_true, 0, 1))
     tp = (K.sum(y_pos * y_pred_pos) + smooth) / (K.sum(y_pos) + smooth)
     return tp
+
 
 def tn(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
@@ -154,18 +171,20 @@ def tn(y_true, y_pred):
     tn = (K.sum(y_neg * y_pred_neg) + smooth) / (K.sum(y_neg) + smooth)
     return tn
 
+
 def focal_tversky(y_true, y_pred):
     # https://github.com/nabsabraham/focal-tversky-unet/blob/master/losses.py
     pt_1 = tversky_loss(y_true, y_pred)
     gamma = 0.75
     return K.pow((pt_1), gamma)
 
+
 def weighted_cat_cross_entropy(y_true, y_pred, class_weights):
     class_weights = tf.reduce_sum(y_true, axis=-1, keepdims=True) / tf.reduce_sum(y_true)
 
-    weights = tf.reduce_sum(class_weights * tf.cast(y_true, 'float64'), axis=-1)
-    unweighted_losses = categorical_crossentropy(tf.cast(y_true,'float32'), tf.cast(y_pred,'float32'))
-    weighted_losses = tf.cast(unweighted_losses,'float32') * tf.cast(weights, 'float32')
+    weights = tf.reduce_sum(class_weights * tf.cast(y_true, tf.float32), axis=-1)
+    unweighted_losses = categorical_crossentropy(tf.cast(y_true, tf.float32), tf.cast(y_pred, tf.float32))
+    weighted_losses = tf.cast(unweighted_losses, tf.float32) * tf.cast(weights, tf.float32)
 
     loss = tf.reduce_mean(weighted_losses)
     return loss
